// src/app/shared/types/cat-breed.ts
export enum CatBreed {
  AmericanCurl = 'american-curl',
  BlackRagdoll = 'black-ragdoll',
  EgyptianMauCat = 'egyptian-mau-cat',
  ExoticShorthair = 'exotic-shorthair',
  JapaneseBobtail = 'japanese-bobtail',
  MaineCoon = 'maine-coon',
  Persian = 'persian',
  ScottishFold = 'scottish-fold',
  SelkirkRexCat = 'selkirk-rex-cat',
  Siamese = 'siamese',
  Singapure = 'singapura',
  TurkishAngora = 'turkish-angora',
}